Added the rest of the account and dashboard rewrite. Which includes the coms to pages.
