+++ client/pages/account/index.vue
@@ -2,81 +2,98 @@
   <div
     class="container w-full px-2 mx-auto my-0 overflow-auto shadow-md md:px-8 md:my-6 bg-gray-50 lg:w-3/5 dark:bg-gray-200"
   >
-    <h1
-      class="py-3 mx-3 mt-3 text-2xl font-bold leading-8 text-center text-gray-900 font-roboto"
-    >
-      Account Settings
-    </h1>
-
     <div class="my-0 sm:my-5">
-      <Profile />
-    </div>
-    <div class="hidden sm:block">
-      <div class="py-5">
-        <div class="border-t border-gray-300"></div>
+      <div class="mt-5 md:mt-0">
+        <h3 class="text-xl font-bold leading-8 text-gray-900">Profile</h3>
+        <p class="text-sm leading-5 text-gray-500">
+          This information will be displayed publicly so be careful what you
+          share.
+        </p>
       </div>
-    </div>
-
-    <div class="my-0 sm:my-5">
-      <PersonalInformation />
-    </div>
-
-    <div class="hidden sm:block">
-      <div class="py-5">
-        <div class="border-t border-gray-300"></div>
+      <div class="mt-6 md:mt-3">
+        <form @submit.prevent="changeProfileInformation">
+          <div class="shadow sm:rounded-md sm:overflow-hidden">
+            <div class="px-4 py-5 bg-white sm:p-6">
+              <label
+                for="username"
+                class="block text-sm font-medium leading-5 text-gray-700"
+              >
+                Username
+              </label>
+              <div class="flex mt-1 rounded-md shadow-sm">
+                <input
+                  id="username"
+                  v-model="changeUsername.username"
+                  :class="{
+                    'border-red-500': changeUsername.errors.username || error,
+                  }"
+                  class="flex-1 block w-full px-3 py-2 mt-1 transition duration-150 ease-in-out border border-gray-300 rounded-md shadow-sm form-input focus:outline-none focus:ring-blue focus:border-blue-300 sm:text-sm sm:leading-5"
+                />
+              </div>
+              <span
+                v-if="changeUsername.errors.username"
+                class="text-red-500"
+                >{{ changeUsername.errors.username }}</span
+              >
+            </div>
+            <div class="px-4 py-3 text-right bg-gray-50 sm:px-6">
+              <span class="inline-flex rounded-md shadow-sm">
+                <button
+                  type="submit"
+                  class="inline-flex justify-center px-4 py-2 text-sm font-medium leading-5 text-white transition duration-150 ease-in-out border border-transparent rounded-md bg-primary-600 hover:bg-primary-500 focus:outline-none focus:border-primary-700"
+                >
+                  Save
+                </button>
+              </span>
+            </div>
+          </div>
+        </form>
       </div>
-    </div>
-    <div class="my-0 sm:my-5">
-      <Security />
-    </div>
-
-    <div class="hidden sm:block">
-      <div class="py-5">
-        <div class="border-t border-gray-300"></div>
-      </div>
-    </div>
-    <div class="my-0 sm:my-5">
-      <Notifications />
-    </div>
-
-    <div class="hidden sm:block">
-      <div class="py-5">
-        <div class="border-t border-gray-300"></div>
-      </div>
-    </div>
-    <div class="my-0 sm:my-5">
-      <DeviceList :sessions="sessions" />
     </div>
   </div>
 </template>
 
 <script>
-import Profile from '@/components/Account/Profile'
-import PersonalInformation from '@/components/Account/PersonalInformation'
-import Security from '@/components/Account/Security'
-import Notifications from '@/components/Account/Notifications'
-import DeviceList from '../../components/Account/DeviceList'
-
 export default {
-  components: {
-    Profile,
-    PersonalInformation,
-    Security,
-    Notifications,
-    DeviceList,
-  },
-
   layout: 'account',
 
   middleware: ['auth'],
 
-  async fetch({ store }) {
-    await store.dispatch('account/FETCH_SESSIONS')
+  data() {
+    return {
+      changeUsername: {
+        username: this.$auth.user.username,
+        errors: { username: null },
+      },
+      error: null,
+      success: null,
+    }
   },
 
-  computed: {
-    sessions() {
-      return this.$store.state.account.sessions
+  methods: {
+    async changeProfileInformation(e) {
+      const changeUsername =
+        this.$auth.user.username !== this.changeUsername.username
+
+      if (changeUsername) {
+        try {
+          const res = await this.$axios.$put('/api/account/change-username', {
+            username: this.changeUsername.username,
+          })
+          await this.$auth.fetchUser()
+          this.$toast.success(res.message, {
+            position: 'bottom-right',
+          })
+        } catch (e) {
+          if (e.response && e.response.data && e.response.data.errors) {
+            this.changeUsername.errors = e.response.data.errors
+          } else {
+            this.$toast.error('Oops.. Something Went Wrong..', {
+              position: 'bottom-right',
+            })
+          }
+        }
+      }
     },
   },
 }
