+++ server/middleware/auth/isRefreshTokenValid.js
@@ -30,6 +30,6 @@ module.exports = async (req, res, next) => {
     res.status(401).send('Unauthorized');
   } catch (err) {
     console.log(err);
-    res.status(500).send('Server error');
+    res.status(500).send('Internal Server Error');
   }
 };
