+++ server/middleware/auth/isSessionValid.js
@@ -35,6 +35,6 @@ module.exports = async (req, res, next) => {
     res.status(401).send('Unauthorized');
   } catch (err) {
     console.log(err);
-    res.status(500).send('Server error');
+    res.status(500).send('Internal Server Error');
   }
 };
