+++ server/middleware/isAccountActivated.js
@@ -21,6 +21,6 @@ module.exports = async (req, res, next) => {
     });
   } catch (err) {
     console.log(err);
-    res.status(500).send('Server error');
+    res.status(500).send('Internal Server Error');
   }
 };
